backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
djangorestframework-simplejwt>=5.3
django-filter>=23.0
Faker>=22.0
